requirements.txt
Django==2.2.8
djangorestframework==3.9.0
pytz==2021.1
sqlparse==0.4.1
